requirements.txt
protobuf==7.34.1
python-dotenv==1.2.2
google-genai==1.69.0